renamed core/simulator to core_legacy_simulator
